+++ src/components/Cards/AccommodationCard.tsx
@@ -6,7 +6,7 @@ import {
   CardActionArea,
 } from "@mui/material";
 import { useTheme } from "@mui/material/styles";
-import { Rating } from "./Rating";
+import { Rating } from "../Rating";
 
 type AccommodationCardProps = {
   title: string;
