+++ src/components/index.ts
@@ -1,14 +1,14 @@
-export {AccommodationCard} from "./AccommodationCard"
+export {AccommodationCard} from "./Cards/AccommodationCard"
 export {AccommodationSearch} from "./AccommodationSearch"
 export {AdvancedSearch} from "./AdvancedSearch"
 export {BookStayButton} from "./BookStayButton"
-export {CityCard} from "./CityCard"
+export {CityCard} from "./Cards/CityCard"
 export {Footer} from "./Footer"
 export {Header} from "./Header"
 export {Navigation} from "./Navigation"
-export {PlaceCard} from "./PlaceCard"
+export {PlaceCard} from "./Cards/PlaceCard"
 export {Rating} from "./Rating"
-export {ReservationCard} from "./ReservationCard"
+export {ReservationCard} from "./Cards/ReservationCard"
 export {ReservationForm} from "./ReservationForm"
 export {SearchButton} from "./SearchButton"
 export {SimpleSearch} from "./SimpleSearch"
